// first/src/app/models/candidat.model.ts
export interface Candidat {
    id_cand: number;
    nomCand: string;
    prenomCand: string;
    emailCand: string;
    dateNaiss: string;
    etat_civil: string;
    sexe: string;
    ville: string;
    codePostal: string;
    numTel: string;
    permis: string;
    niveauEducation: string;
    diplomeObtenu: string;
    etablissement: string;
    niveauExperience: string;
    statutActuel: string;
    langues: string;
    competences: string;
    posteDesire: string;
    lieuTravailSouhaite: string;
  }
  